# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AYLink Agent 的 HTTP API 文档。",
        "title": "AYLink Agent API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/control/webrtc-network": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回 WebRTC ICE 策略和 ICE server 设置。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "获取 WebRTC 网络设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.WebRtcNetworkSettings"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/i18n/languages": {
            "get": {
                "description": "返回 Agent 支持的语言列表。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "国际化"
                ],
                "summary": "获取语言列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.LanguageOption"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/i18n/{locale}": {
            "get": {
                "description": "返回指定语言区域的语言包，不存在时回退到服务端语言设置。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "国际化"
                ],
                "summary": "获取语言包",
                "parameters": [
                    {
                        "type": "string",
                        "description": "语言区域代码",
                        "name": "locale",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/settings/language": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前服务端语言设置。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "获取服务端语言",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ServerLanguageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "保存服务端语言设置。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "保存服务端语言",
                "parameters": [
                    {
                        "description": "语言设置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ServerLanguageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ServerLanguageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings/webrtc-network": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回 WebRTC ICE 策略和 ICE server 设置。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "获取 WebRTC 网络设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.WebRtcNetworkSettings"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "保存 WebRTC ICE 策略和 ICE server 设置。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "保存 WebRTC 网络设置",
                "parameters": [
                    {
                        "description": "WebRTC 网络设置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.WebRtcNetworkSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.WebRtcNetworkSettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "messageKey": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handler.ErrorPayload"
                }
            }
        },
...
        "handler.LanguageOption": {
            "type": "object",
            "properties": {
                "locale": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handler.ServerLanguageRequest": {
            "type": "object",
            "properties": {
                "locale": {
                    "type": "string"
                }
            }
        },
        "handler.ServerLanguageResponse": {
            "type": "object",
            "properties": {
                "locale": {
                    "type": "string"
                }
            }
        },
...
        "handler.WebRtcNetworkSettings": {
            "type": "object",
            "properties": {
                "FallbackLocale": {
                    "type": "string"
                },
                "HostCandidateOverrideEnabled": {
                    "type": "boolean"
                },
                "HostCandidateOverrideIPs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "HostCandidatePortMax": {
                    "type": "integer"
                },
                "HostCandidatePortMin": {
                    "type": "integer"
                },
                "IceServers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/settings.WebRtcIceServer"
                    }
                },
                "IceTransportPolicy": {
                    "type": "string"
                },
                "SinglePortMuxBindPort": {
                    "type": "integer"
                },
                "SinglePortMuxEnabled": {
                    "type": "boolean"
                },
                "SinglePortMuxPublishPort": {
                    "type": "integer"
                }
            }
        },
        "settings.WebRtcIceServer": {
            "type": "object",
            "properties": {
                "Credential": {
                    "type": "string"
                },
                "Urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer 访问令牌，例如：Bearer {accessToken}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}